[refactor] Clean the module
[refactor] Updates module docstring

It is a more clear description that fit better with PEP 257 norms.

[fix] Fix Optional from typing uses

[refactor] Move code in pyimgui_utils module

Previously, the source code was at root level. It wasn't recognized as a module. This commit creates a folder to contains the source code. Therefore, the folder is a module: pyimgui_utils.
Additionally, add typing_extensions as a requirement.txt and remove unused import in window.py. Plus, fix typo in requirement.txt.

[refactor] Finetune hold default color btn and alter btn id push scope

Default value of hold color is grey.

[refactor] Refine type hint in component.py

Union[_, None] are now Optional[_]. Also, replace list for List type in function arguments hint.

[refactor] Make window_focus_state a class attribute

Previously, _WINDOWS_FOCUS_STATE was a global attribute in the file. ImGuiWindowAbstract accessed it without holding it. Now it is a class attribute.
Additionally, focus_id attribute is now unprotected.

[chore] Remove unused attributes
Additionally, make draw_content unprotected.